fixed packages compatibility

+++ AMDock/qradialbar.py
@@ -78,7 +78,7 @@ class QRadialBar(QWidget):
         pen.setColor(self._ProgressColor)
         valueAngle = float(float(self._value - self.min_value)/float(self.max_value - self.min_value)) * float(spanAngle)  #Map _value to angle range
         painter.setPen(pen)
-        painter.drawArc(rect.adjusted(offset, offset, -offset, -offset), startAngle * 16, valueAngle * 16)
+        painter.drawArc(rect.adjusted(offset, offset, -offset, -offset), int(startAngle * 16), int(valueAngle * 16))
         painter.restore()
 
     @pyqtSlot(float)
